feat(cluster): add failureDomain spec label
The goal of this commit is to allow capz users to specify
which failure domains are eligible for control plane rollouts.

There's a new label in AzureCluster.spec.failureDomain that is used
to post-filter discovered failure domains.

The field is optional - if it's missing, all discovered failure domains
are eligible for control plane rollout.

THERE IS NO BREAKING CHANGE TO CURRENT USERS.